pyproject.toml
[project]
name = "str-buffet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.6",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "streamlit>=1.49.1",
    "yahooquery>=2.4.1",
]
